refactor(collab): code optimisation and bug fixes
