+++ src/visualBuilder/components/Collab/ThreadPopup/CommentTextArea.tsx
@@ -7,7 +7,7 @@ import {
     useCallback,
     useRef,
 } from "preact/hooks";
-import { collabStyles } from "../../../visualBuilder.style";
+import { collabStyles } from "../../../collab.style";
 import {
     ICommentPayload,
     ICommentResponse,
@@ -34,6 +34,7 @@ import classNames from "classnames";
 
 interface ICommentTextArea {
     userState: IUserState;
+    onClose: (isResolved?: boolean) => void;
     handleOnSaveRef: React.MutableRefObject<any>;
     comment?: IMessageDTO | null;
 }
@@ -47,7 +48,7 @@ const initialState: ICommentState = {
 };
 
 const CommentTextArea: React.FC<ICommentTextArea> = React.memo(
-    ({ userState, handleOnSaveRef, comment }) => {
+    ({ userState, handleOnSaveRef, comment, onClose }) => {
         const [state, setState] = useState<ICommentState>(initialState);
         const [showSuggestions, setShowSuggestions] = useState(false);
         const [cursorPosition, setCursorPosition] = useState({
@@ -73,6 +74,68 @@ const CommentTextArea: React.FC<ICommentTextArea> = React.memo(
             ".collab-thread-body--input--textarea",
             state.message
         );
+
+        useEffect(() => {
+            const textArea = document.getElementById(
+                "collab-thread-body--input--textarea"
+            );
+            if (!textArea) return;
+
+            const baseClasses = {
+                focus: {
+                    base: "collab-thread-body--input--textarea--focus",
+                    goober: collabStyles()[
+                        "collab-thread-body--input--textarea--focus"
+                    ],
+                },
+                hover: {
+                    base: "collab-thread-body--input--textarea--hover",
+                    goober: collabStyles()[
+                        "collab-thread-body--input--textarea--hover"
+                    ],
+                },
+            };
+
+            const handleFocus = () => {
+                textArea.classList.add(
+                    baseClasses.focus.base,
+                    baseClasses.focus.goober
+                );
+            };
+
+            const handleBlur = () => {
+                textArea.classList.remove(
+                    baseClasses.focus.base,
+                    baseClasses.focus.goober
+                );
+            };
+
+            const handleMouseEnter = () => {
+                textArea.classList.add(
+                    baseClasses.hover.base,
+                    baseClasses.hover.goober
+                );
+            };
+
+            const handleMouseLeave = () => {
+                textArea.classList.remove(
+                    baseClasses.hover.base,
+                    baseClasses.hover.goober
+                );
+            };
+
+            textArea.addEventListener("focus", handleFocus);
+            textArea.addEventListener("blur", handleBlur);
+            textArea.addEventListener("mouseenter", handleMouseEnter);
+            textArea.addEventListener("mouseleave", handleMouseLeave);
+
+            return () => {
+                textArea.removeEventListener("focus", handleFocus);
+                textArea.removeEventListener("blur", handleBlur);
+                textArea.removeEventListener("mouseenter", handleMouseEnter);
+                textArea.removeEventListener("mouseleave", handleMouseLeave);
+            };
+        }, []);
 
         // Find the @ symbol position and extract search term
         const findMentionSearchPosition = (text: string, cursorPos: any) => {
@@ -282,6 +345,7 @@ const CommentTextArea: React.FC<ICommentTextArea> = React.memo(
                             comments: updatedComments, // Update the comments list in the state
                         };
                     });
+                    onClose(false);
                 } else {
                     // If creating a new comment, call the create function and add the new comment to the state.
                     let commentResponse: ICommentResponse =
@@ -296,6 +360,7 @@ const CommentTextArea: React.FC<ICommentTextArea> = React.memo(
                     }));
                     // Reset the state to its initial form after a successful comment creation.
                     setState(initialState);
+                    onClose(false);
                 }
             } catch (error: any) {}
         }, [error.hasError, state, activeThread]);
