+++ src/visualBuilder/hooks/useCollabIndicator.ts
@@ -3,6 +3,8 @@ import React from "preact/compat";
 import { useState, useRef, useEffect } from "preact/hooks";
 import { IActiveThread } from "../types/collab.types";
 import Config from "../../configManager/configManager";
+import { calculatePopupPosition } from "../generators/generateThread";
+import { handleEmptyThreads } from "../generators/generateThread";
 
 interface UseCollabPopupProps {
     newThread?: boolean;
@@ -23,45 +25,44 @@ export const useCollabIndicator = ({
         return thread || { _id: "new" };
     });
 
-    const calculatePopupPosition = () => {
-        if (!buttonRef.current || !popupRef.current) return;
-
-        const buttonRect = buttonRef.current.getBoundingClientRect();
-        const popupHeight = 422;
-        const popupWidth = 334;
-        const viewportHeight = window.innerHeight;
-        const viewportWidth = window.innerWidth;
-
-        const spaceAbove = buttonRect.top;
-        const spaceBelow = viewportHeight - buttonRect.bottom;
-
-        let top, left;
-
-        if (spaceAbove >= popupHeight) {
-            top = buttonRect.top - popupHeight - 8;
-        } else {
-            top = buttonRect.bottom + 8;
+    const updatePopupPosition = () => {
+        if (buttonRef.current && popupRef.current) {
+            calculatePopupPosition(buttonRef.current, popupRef.current);
         }
-
-        left = buttonRect.left + buttonRect.width / 2 - popupWidth / 2;
-
-        top = Math.max(top, 0);
-        left = Math.max(left, 0);
-        left = Math.min(left, viewportWidth - popupWidth);
-
-        popupRef.current.style.top = `${top}px`;
-        popupRef.current.style.left = `${left}px`;
     };
 
     useEffect(() => {
         if (!showPopup) return;
-        calculatePopupPosition();
+        updatePopupPosition();
     }, [showPopup]);
+
+    useEffect(() => {
+        const handleExternalClose = () => {
+            handleEmptyThreads();
+            const closestDiv = buttonRef.current?.closest("div[field-path]");
+            if (closestDiv) {
+                (closestDiv as HTMLElement).style.zIndex = "999";
+            }
+            setShowPopup(false);
+        };
+
+        document.addEventListener("closeCollabPopup", handleExternalClose);
+        return () => {
+            document.removeEventListener(
+                "closeCollabPopup",
+                handleExternalClose
+            );
+        };
+    }, []);
 
     const togglePopup = () => {
         if (!showPopup) {
             document.dispatchEvent(new CustomEvent("closeCollabPopup"));
             setShowPopup(true);
+            const closestDiv = buttonRef.current?.closest("div[field-path]");
+            if (closestDiv) {
+                (closestDiv as HTMLElement).style.zIndex = "1000";
+            }
         } else {
             setShowPopup(false);
             if (config?.collab?.isFeedbackMode === false) {
