+++ src/visualBuilder/index.ts
@@ -46,7 +46,10 @@ import { updatePopupPositions } from "./generators/generateThread";
 import { useRecalculateVariantDataCSLPValues } from "./eventManager/useRecalculateVariantDataCSLPValues";
 import { VB_EmptyBlockParentClass } from "..";
 import { useCollab } from "./eventManager/useCollab";
-import { generateThreadsFromData } from "./generators/generateThread";
+import {
+    generateThread,
+    handleMissingThreads,
+} from "./generators/generateThread";
 import { IThreadDTO } from "./types/collab.types";
 
 interface VisualBuilderGlobalStateImpl {
@@ -203,8 +206,21 @@ export class VisualBuilder {
                     ".visual-builder__container"
                 );
                 if (container && threadsPayload) {
-                    generateThreadsFromData(threadsPayload);
+                    const missingThreadIds = threadsPayload
+                        ?.map((payload: IThreadDTO) =>
+                            generateThread(payload, { isNewThread: false })
+                        )
+                        .filter(
+                            (threadId): threadId is string =>
+                                threadId !== undefined
+                        );
                     threadsPayload = [];
+                    if (missingThreadIds.length > 0) {
+                        handleMissingThreads({
+                            payload: { isElementPresent: false },
+                            threadUids: missingThreadIds,
+                        });
+                    }
                 }
 
                 const emptyBlockParents = Array.from(
