+++ src/visualBuilder/utils/types/index.types.ts
@@ -54,7 +54,6 @@ export enum FieldDataType {
     FILE = "file",
     GLOBAL_FIELD = "global_field",
     TAXONOMY = "taxonomy",
-    DISCUSSION = "discussion",
 }
 
 export interface VisualBuilderEditContext {
